// src/app/dto/sentiment-per-library.dto.ts
export interface SentimentPerLibrary {
  vader_very_positive: number;
  vader_positive: number;
  vader_neutral: number;
  vader_negative: number;
  vader_very_negative: number;
  pattern_very_positive: number;
  pattern_positive: number;
  pattern_neutral: number;
  pattern_negative: number;
  pattern_very_negative: number;
  text_blob_very_positive: number;
  text_blob_positive: number;
  text_blob_neutral: number;
  text_blob_negative: number;
  text_blob_very_negative: number;
}